add emails

+++ app/Http/Controllers/JobController.php
@@ -3,11 +3,13 @@
 namespace App\Http\Controllers;
 
 
+use App\Mail\JobPosted;
 use App\Models\Job;
 use App\Models\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
@@ -33,11 +35,15 @@ class JobController extends Controller
             'salary' => ['required'],
         ]);
 
-        Job::create([
+        $job = Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
+
+        Mail::to($job->employer->user)->send(
+            new JobPosted($job)
+        );
 
         return redirect('/jobs');
     }
@@ -50,12 +56,14 @@ class JobController extends Controller
     }
 
     public function update(Job $job) {
+
+        Gate::authorize('edit-job', $job);
+
         request()->validate([
             'title' => ['required', 'min:3'],
             'salary' => ['required'],
         ]);
 
-        // todo authorize
 
         $job->update([
             'title' => request('title'),
@@ -66,7 +74,7 @@ class JobController extends Controller
     }
 
     public function destroy(Job $job) {
-    // todo authorize
+        Gate::authorize('edit-job', $job);
 
         $job->delete();
